update SimulatedPlayerCoordinator navmesh
