+++ workers/unity/Assets/Fps/Scenes/FPS-SimulatedPlayerCoordinator.unity
@@ -125,7 +125,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 1613497914398450, guid: e4a3eb8590344604db1e998db676a211, type: 3}
       propertyPath: m_IsActive
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 4593588597980260, guid: e4a3eb8590344604db1e998db676a211, type: 3}
       propertyPath: m_LocalPosition.x
