+++ workers/unity/Assets/Fps/Scripts/SetupLogic/AndroidWorkerConnector.cs
@@ -14,28 +14,9 @@ using Improbable.Gdk.Mobile.Android;
 namespace Fps
 {
     [RequireComponent(typeof(ConnectionController))]
-    public class AndroidWorkerConnector : MobileWorkerConnector, ITileProvider
+    public class AndroidWorkerConnector : MobileWorkerConnectorBase, ITileProvider
     {
-        private const string AuthPlayer = "Prefabs/MobileClient/Authoritative/Player";
-        private const string NonAuthPlayer = "Prefabs/MobileClient/NonAuthoritative/Player";
-
-        private const string Small = "small";
-        private const string Large = "large";
-
         public bool ShouldConnectLocally;
-        public int TargetFrameRate = 60;
-
-        public GameObject SmallLevelPrefab;
-        public GameObject LargeLevelPrefab;
-
-        private GameObject levelInstance;
-
-        private List<TileEnabler> levelTiles = new List<TileEnabler>();
-        public List<TileEnabler> LevelTiles => levelTiles;
-
-        private ConnectionController connectionController;
-
-        public string IpAddress { get; set; }
 
         private void Awake()
         {
@@ -53,6 +34,11 @@ namespace Fps
                     Debug.LogWarning("Unable to find DevAuthToken.txt in the Resources folder.");
                 }
             }
+        }
+
+        protected override string GetWorkerType()
+        {
+            return WorkerUtils.AndroidClient;
         }
 
         protected virtual async void Start()
@@ -109,12 +95,6 @@ namespace Fps
             await Connect(WorkerUtils.AndroidClient, new ForwardingDispatcher()).ConfigureAwait(false);
         }
 
-        protected override string SelectDeploymentName(DeploymentList deployments)
-        {
-            // This could be replaced with a splash screen asking to select a deployment or some other user-defined logic.
-            return deployments.Deployments[0].DeploymentName;
-        }
-
         protected override ConnectionService GetConnectionService()
         {
             if (ShouldConnectLocally)
@@ -123,73 +103,6 @@ namespace Fps
             }
 
             return ConnectionService.AlphaLocator;
-        }
-
-        protected override void HandleWorkerConnectionEstablished()
-        {
-            var world = Worker.World;
-
-            // Only take the Heartbeat from the PlayerLifecycleConfig Client Systems.
-            world.GetOrCreateManager<HandlePlayerHeartbeatRequestSystem>();
-
-            GameObjectRepresentationHelper.AddSystems(world);
-            var fallback = new GameObjectCreatorFromMetadata(Worker.WorkerType, Worker.Origin, Worker.LogDispatcher);
-
-            // Set the Worker gameObject to the ClientWorker so it can access PlayerCreater reader/writers
-            GameObjectCreationHelper.EnableStandardGameObjectCreation(
-                world,
-                new AdvancedEntityPipeline(Worker, AuthPlayer, NonAuthPlayer, fallback),
-                gameObject);
-
-            LoadWorld();
-        }
-
-        protected override void HandleWorkerConnectionFailure(string errorMessage)
-        {
-            connectionController.OnFailedToConnect();
-        }
-
-        public override void Dispose()
-        {
-            if (levelInstance != null)
-            {
-                Destroy(levelInstance);
-            }
-
-            base.Dispose();
-        }
-
-        // Get the world size from the config, and use it to load the appropriate level.
-        protected virtual void LoadWorld()
-        {
-            var workerSystem = Worker.World.GetExistingManager<WorkerSystem>();
-            var worldSize = workerSystem.Connection.GetWorkerFlag("world_size");
-
-            if (worldSize != Small && worldSize != Large)
-            {
-                workerSystem.LogDispatcher.HandleLog(LogType.Error,
-                    new LogEvent(
-                            "Invalid world_size worker flag. Make sure that it is either small, medium, or large,")
-                        .WithField("world_size", worldSize));
-                return;
-            }
-
-            var levelToLoad = worldSize == Large ? LargeLevelPrefab : SmallLevelPrefab;
-
-            if (levelToLoad == null)
-            {
-                Debug.LogError("The level to be instantiated is null.");
-                return;
-            }
-
-            levelInstance = Instantiate(levelToLoad, transform.position, transform.rotation);
-            levelInstance.name = $"{levelToLoad.name}({Worker.WorkerType})";
-
-            levelInstance.GetComponentsInChildren<TileEnabler>(true, levelTiles);
-            foreach (var tileEnabler in levelTiles)
-            {
-                tileEnabler.IsClient = true;
-            }
         }
     }
 }
