+++ workers/unity/Assets/Fps/Scripts/SetupLogic/MobileWorkerConnectorBase.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d7f4de6841404b9f8db5e70bc303c84f
+timeCreated: 1549970379
